moved string utility functions from util/met_util to util/string, ci-run-all-diff
