+++ metplus/util/__init__.py
@@ -1,4 +1,5 @@
 from .constants import *
+from .string import *
 from .metplus_check import *
 from .doc_util import *
 from .config_metplus import *
