+++ metplus/util/met_config.py
@@ -5,7 +5,8 @@ Contact(s): George McCabe
 
 import os
 
-from .met_util import getlist, get_threshold_via_regex, MISSING_DATA_VALUE
+from .string import getlist
+from .met_util import get_threshold_via_regex, MISSING_DATA_VALUE
 from .met_util import remove_quotes as util_remove_quotes
 from .config_metplus import find_indices_in_config_section
