+++ metplus/wrappers/tc_pairs_wrapper.py
@@ -20,6 +20,7 @@ import csv
 import datetime
 import glob
 
+from ..util import getlist
 from ..util import time_util
 from ..util import met_util as util
 from ..util import do_string_sub
@@ -170,10 +171,10 @@ class TCPairsWrapper(CommandBuilder):
         # if looping by processes, get the init or valid beg time and run once
         c_dict['INPUT_DICT'] = self.get_start_time_input_dict()
 
-        c_dict['INIT_INCLUDE'] = util.getlist(
+        c_dict['INIT_INCLUDE'] = getlist(
             self.get_wrapper_or_generic_config('INIT_INCLUDE')
         )
-        c_dict['INIT_EXCLUDE'] = util.getlist(
+        c_dict['INIT_EXCLUDE'] = getlist(
             self.get_wrapper_or_generic_config('INIT_EXCLUDE')
         )
         c_dict['VALID_BEG'] = self.get_wrapper_or_generic_config('VALID_BEG')
@@ -195,7 +196,7 @@ class TCPairsWrapper(CommandBuilder):
         )
 
         # get list of models to process
-        c_dict['MODEL_LIST'] = util.getlist(
+        c_dict['MODEL_LIST'] = getlist(
             self.config.getraw('config', 'MODEL', '')
         )
         # if no models are requested, set list to contain a single string
@@ -205,7 +206,7 @@ class TCPairsWrapper(CommandBuilder):
 
         self._read_storm_info(c_dict)
 
-        c_dict['STORM_NAME_LIST'] = util.getlist(
+        c_dict['STORM_NAME_LIST'] = getlist(
             self.config.getraw('config', 'TC_PAIRS_STORM_NAME')
         )
         c_dict['DLAND_FILE'] = self.config.getraw('config',
@@ -287,13 +288,13 @@ class TCPairsWrapper(CommandBuilder):
         @param c_dict dictionary to populate with values from config
         @returns None
         """
-        storm_id_list = util.getlist(
+        storm_id_list = getlist(
             self.config.getraw('config', 'TC_PAIRS_STORM_ID', '')
         )
-        cyclone_list = util.getlist(
+        cyclone_list = getlist(
             self.config.getraw('config', 'TC_PAIRS_CYCLONE', '')
         )
-        basin_list = util.getlist(
+        basin_list = getlist(
             self.config.getraw('config', 'TC_PAIRS_BASIN', '')
         )
